requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.5
django-admin-thumbnails==0.2.8
django-appconf==1.0.5
django-background-task==0.1.8
django-compat==1.0.15
django-cors-headers==4.1.0
django-crontab==0.7.1
django-daraja==1.3.0
django-imagekit==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openapi-codec==1.3.2
pilkit==2.0
Pillow==10.0.1
pycparser==2.21
Pygments==2.15.1
PyJWT==1.7.1
PyMySQL==1.1.0
PySocks==1.7.1
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.2
